turn on retire only when needed

+++ concurrency_control/row_clv.cpp
@@ -129,9 +129,9 @@ RC Row_clv::lock_get(lock_t type, txn_man * txn, uint64_t* &txnids, int &txncnt)
 	insert_to_waiters(type, txn);
 
 	// turn on retire only when needed
-	if (!retire_on && waiter_cnt > min(g_thread_cnt / 2, 4))
+	if (!retire_on && waiter_cnt > min(g_thread_cnt / 2, (UInt32) 4))
 		retire_on = true;
-	else if ( retired_cnt >= max(g_thread_cnt / 2, 2) ){
+	else if ( retired_cnt >= max(g_thread_cnt / 2, (UInt32) 2) ){
 		retire_on = false;
 	}
